+++ test/examples_tests.dart
@@ -32,13 +32,17 @@ class ExamplesTests
         await wordsApi.acceptAllRevisions(request);
     }
 
-    Future<void> acceptAllRevisions2() async
+    Future<void> acceptAllRevisionsOnline() async
     {
         final wordsApi = WordsApi(config);
         final fileName  = 'test_doc.docx';
 
-        // Calls AcceptAllRevisions method for document in cloud.
-        final request = AcceptAllRevisionsRequest(fileName);
-        await wordsApi.acceptAllRevisions(request);
+        // Calls AcceptAllRevisionsOnline method for document in cloud.
+        final documentData = (await File(fileName).readAsBytes()).buffer.asByteData();
+        final request = AcceptAllRevisionsOnlineRequest(documentData);
+        final acceptAllRevisionsOnlineResult = await wordsApi.acceptAllRevisionsOnline(request);
+        await File('test_result.docx').writeAsBytes(
+            acceptAllRevisionsOnlineResult.document.buffer.asUint8List(acceptAllRevisionsOnlineResult.document.offsetInBytes, acceptAllRevisionsOnlineResult.document.lengthInBytes)
+        );
     }
 }